+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 1.1.1 (UNRELEASED)
 ------------------
 
+* #38: Reintroduce ``extra_args`` argument to ``Callback._GetKey``, so subclasses can make use
+  of it.
 * #36: Fix regression introduced in ``1.1.0`` where ``Abstract`` and ``Implements`` decorators
   could no longer be used in interfaces implementations.